refactor(Events): 공통 이벤트 패키지 분리 (#89)
이벤트 발행 인프라는 allreva.events로 이동
KeywordSearchedEvent는 search 도메인 이벤트로 옮겨 common 의존을 줄임
