+++ src/main/java/com/backend/allreva/events/Event.java
@@ -1,4 +1,4 @@
-package com.backend.allreva.common.event;
+package com.backend.allreva.events;
 
 import lombok.Getter;
