+++ src/main/java/com/backend/allreva/module/notification/infra/sse/event/SseTimedOutEvent.java
@@ -1,6 +1,6 @@
 package com.backend.allreva.module.notification.infra.sse.event;
 
-import com.backend.allreva.common.event.Event;
+import com.backend.allreva.events.Event;
 import lombok.Getter;
 
 @Getter
